refactored directory / file names KIN-1046
