+++ packages/common/src/components/attribute_selectors/IconSelect.js
