+++ packages/common/src/components/attribute_selectors/NotificationTemplateSelect.js
