+++ packages/common/src/components/attribute_selectors/fa-icons.json
