+++ packages/discussions/src/index.js
@@ -7,7 +7,7 @@ import {
   UploadMessage,
   MessagesGroup,
 } from './components/MessagesGroup';
-import * as DiscussionAPI from './discussion_api';
+import * as DiscussionAPI from './discussionApi';
 
 // Components
 export { Discussion, TextMessage, UploadMessage, MessagesGroup };
